Update styles and refactor app component structure
Updated the application's CSS to import a new theme and added the primeflex library. The app component HTML was refactored to utilize new custom components, improving modularity and readability. Furthermore, updates were implemented to the app component TypeScript to accommodate the new component structure and additional information displayed. The package-lock.json file was automatically updated with the new library addition.